# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ethiopian Restaurant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MenuItem {
  id: number;
  name: string;
  /** @nullable */
  nameAmharic?: string | null;
  description: string;
  price: number;
  category: string;
  isVegan: boolean;
  isVegetarian: boolean;
  isGlutenFree: boolean;
  isAvailable: boolean;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateMenuItemBody {
  name: string;
  /** @nullable */
  nameAmharic?: string | null;
  description: string;
  price: number;
  category: string;
  isVegan?: boolean;
  isVegetarian?: boolean;
  isGlutenFree?: boolean;
  isAvailable?: boolean;
  /** @nullable */
  imageUrl?: string | null;
}

export interface UpdateMenuItemBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  nameAmharic?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  isVegan?: boolean | null;
  /** @nullable */
  isVegetarian?: boolean | null;
  /** @nullable */
  isGlutenFree?: boolean | null;
  /** @nullable */
  isAvailable?: boolean | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface MenuCategory {
  category: string;
  itemCount: number;
}

export interface DailyMenuItem {
  id: number;
  date: string;
  name: string;
  description: string;
  price: number;
  isVegan: boolean;
  isVegetarian: boolean;
  /** @nullable */
  note?: string | null;
  createdAt: string;
}

export interface CreateDailyMenuItemBody {
  /** @nullable */
  date?: string | null;
  name: string;
  description: string;
  price: number;
  isVegan?: boolean;
  isVegetarian?: boolean;
  /** @nullable */
  note?: string | null;
}

export interface TodaysDailyMenu {
  date: string;
  items: DailyMenuItem[];
  totalItems: number;
  veganCount: number;
  vegetarianCount: number;
}

export interface SocialPost {
  id: number;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  platforms: string[];
  status: string;
  /** @nullable */
  publishedAt?: string | null;
  createdAt: string;
}

export interface CreateSocialPostBody {
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  platforms: string[];
  publishNow?: boolean;
}

export interface SocialPostStats {
  totalPosts: number;
  publishedPosts: number;
  draftPosts: number;
  instagramPosts: number;
  facebookPosts: number;
  tiktokPosts: number;
}

export interface GalleryImage {
  id: number;
  url: string;
  /** @nullable */
  caption?: string | null;
  category: string;
  sortOrder: number;
  createdAt: string;
}

export interface CreateGalleryImageBody {
  url: string;
  /** @nullable */
  caption?: string | null;
  category: string;
  sortOrder?: number;
}

export type GetDailyMenuItemsParams = {
  /**
   * @nullable
   */
  date?: string | null;
};