# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ethiopian Restaurant API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get all menu items
 */
export const GetMenuItemsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  nameAmharic: zod.string().nullish(),
  description: zod.string(),
  price: zod.number(),
  category: zod.string(),
  isVegan: zod.boolean(),
  isVegetarian: zod.boolean(),
  isGlutenFree: zod.boolean(),
  isAvailable: zod.boolean(),
  imageUrl: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const GetMenuItemsResponse = zod.array(GetMenuItemsResponseItem);

/**
 * @summary Create a menu item
 */
export const CreateMenuItemBody = zod.object({
  name: zod.string(),
  nameAmharic: zod.string().nullish(),
  description: zod.string(),
  price: zod.number(),
  category: zod.string(),
  isVegan: zod.boolean().optional(),
  isVegetarian: zod.boolean().optional(),
  isGlutenFree: zod.boolean().optional(),
  isAvailable: zod.boolean().optional(),
  imageUrl: zod.string().nullish(),
});

/**
 * @summary Update a menu item
 */
export const UpdateMenuItemParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateMenuItemBody = zod.object({
  name: zod.string().nullish(),
  nameAmharic: zod.string().nullish(),
  description: zod.string().nullish(),
  price: zod.number().nullish(),
  category: zod.string().nullish(),
  isVegan: zod.boolean().nullish(),
  isVegetarian: zod.boolean().nullish(),
  isGlutenFree: zod.boolean().nullish(),
  isAvailable: zod.boolean().nullish(),
  imageUrl: zod.string().nullish(),
});

export const UpdateMenuItemResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  nameAmharic: zod.string().nullish(),
  description: zod.string(),
  price: zod.number(),
  category: zod.string(),
  isVegan: zod.boolean(),
  isVegetarian: zod.boolean(),
  isGlutenFree: zod.boolean(),
  isAvailable: zod.boolean(),
  imageUrl: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a menu item
 */
export const DeleteMenuItemParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get all menu categories with item counts
 */
export const GetMenuCategoriesResponseItem = zod.object({
  category: zod.string(),
  itemCount: zod.number(),
});
export const GetMenuCategoriesResponse = zod.array(
  GetMenuCategoriesResponseItem,
);

/**
 * @summary Get today's daily menu
 */
export const GetDailyMenuItemsQueryParams = zod.object({
  date: zod.coerce.string().nullish(),
});

export const GetDailyMenuItemsResponseItem = zod.object({
  id: zod.number(),
  date: zod.string(),
  name: zod.string(),
  description: zod.string(),
  price: zod.number(),
  isVegan: zod.boolean(),
  isVegetarian: zod.boolean(),
  note: zod.string().nullish(),
  createdAt: zod.string(),
});
export const GetDailyMenuItemsResponse = zod.array(
  GetDailyMenuItemsResponseItem,
);

/**
 * @summary Add item to daily menu
 */
export const CreateDailyMenuItemBody = zod.object({
  date: zod.string().nullish(),
  name: zod.string(),
  description: zod.string(),
  price: zod.number(),
  isVegan: zod.boolean().optional(),
  isVegetarian: zod.boolean().optional(),
  note: zod.string().nullish(),
});

/**
 * @summary Remove item from daily menu
 */
export const DeleteDailyMenuItemParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get today's complete daily menu summary
 */
export const GetTodaysDailyMenuResponse = zod.object({
  date: zod.string(),
  items: zod.array(
    zod.object({
      id: zod.number(),
      date: zod.string(),
      name: zod.string(),
      description: zod.string(),
      price: zod.number(),
      isVegan: zod.boolean(),
      isVegetarian: zod.boolean(),
      note: zod.string().nullish(),
      createdAt: zod.string(),
    }),
  ),
  totalItems: zod.number(),
  veganCount: zod.number(),
  vegetarianCount: zod.number(),
});

/**
 * @summary Get all social media posts
 */
export const GetSocialPostsResponseItem = zod.object({
  id: zod.number(),
  content: zod.string(),
  imageUrl: zod.string().nullish(),
  platforms: zod.array(zod.string()),
  status: zod.string(),
  publishedAt: zod.string().nullish(),
  createdAt: zod.string(),
});
export const GetSocialPostsResponse = zod.array(GetSocialPostsResponseItem);

/**
 * @summary Create and publish a social media post
 */
export const CreateSocialPostBody = zod.object({
  content: zod.string(),
  imageUrl: zod.string().nullish(),
  platforms: zod.array(zod.string()),
  publishNow: zod.boolean().optional(),
});

/**
 * @summary Delete a social media post
 */
export const DeleteSocialPostParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get social media post statistics
 */
export const GetSocialPostStatsResponse = zod.object({
  totalPosts: zod.number(),
  publishedPosts: zod.number(),
  draftPosts: zod.number(),
  instagramPosts: zod.number(),
  facebookPosts: zod.number(),
  tiktokPosts: zod.number(),
});

/**
 * @summary Get gallery images
 */
export const GetGalleryImagesResponseItem = zod.object({
  id: zod.number(),
  url: zod.string(),
  caption: zod.string().nullish(),
  category: zod.string(),
  sortOrder: zod.number(),
  createdAt: zod.string(),
});
export const GetGalleryImagesResponse = zod.array(GetGalleryImagesResponseItem);

/**
 * @summary Add gallery image
 */
export const CreateGalleryImageBody = zod.object({
  url: zod.string(),
  caption: zod.string().nullish(),
  category: zod.string(),
  sortOrder: zod.number().optional(),
});