# lib/api-zod/src/generated/types/createDailyMenuItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ethiopian Restaurant API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateDailyMenuItemBody {
  /** @nullable */
  date?: string | null;
  name: string;
  description: string;
  price: number;
  isVegan?: boolean;
  isVegetarian?: boolean;
  /** @nullable */
  note?: string | null;
}