# lib/api-zod/src/generated/types/createGalleryImageBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ethiopian Restaurant API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateGalleryImageBody {
  url: string;
  /** @nullable */
  caption?: string | null;
  category: string;
  sortOrder?: number;
}