# lib/api-zod/src/generated/types/createMenuItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ethiopian Restaurant API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMenuItemBody {
  name: string;
  /** @nullable */
  nameAmharic?: string | null;
  description: string;
  price: number;
  category: string;
  isVegan?: boolean;
  isVegetarian?: boolean;
  isGlutenFree?: boolean;
  isAvailable?: boolean;
  /** @nullable */
  imageUrl?: string | null;
}