# lib/api-zod/src/generated/types/createSocialPostBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ethiopian Restaurant API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSocialPostBody {
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  platforms: string[];
  publishNow?: boolean;
}