# lib/api-zod/src/generated/types/dailyMenuItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ethiopian Restaurant API
 * OpenAPI spec version: 0.1.0
 */

export interface DailyMenuItem {
  id: number;
  date: string;
  name: string;
  description: string;
  price: number;
  isVegan: boolean;
  isVegetarian: boolean;
  /** @nullable */
  note?: string | null;
  createdAt: string;
}