# lib/api-zod/src/generated/types/menuCategory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ethiopian Restaurant API
 * OpenAPI spec version: 0.1.0
 */

export interface MenuCategory {
  category: string;
  itemCount: number;
}