# lib/api-zod/src/generated/types/socialPost.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ethiopian Restaurant API
 * OpenAPI spec version: 0.1.0
 */

export interface SocialPost {
  id: number;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  platforms: string[];
  status: string;
  /** @nullable */
  publishedAt?: string | null;
  createdAt: string;
}