# lib/api-zod/src/generated/types/socialPostStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ethiopian Restaurant API
 * OpenAPI spec version: 0.1.0
 */

export interface SocialPostStats {
  totalPosts: number;
  publishedPosts: number;
  draftPosts: number;
  instagramPosts: number;
  facebookPosts: number;
  tiktokPosts: number;
}