# lib/api-zod/src/generated/types/todaysDailyMenu.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ethiopian Restaurant API
 * OpenAPI spec version: 0.1.0
 */
import type { DailyMenuItem } from "./dailyMenuItem";

export interface TodaysDailyMenu {
  date: string;
  items: DailyMenuItem[];
  totalItems: number;
  veganCount: number;
  vegetarianCount: number;
}