# lib/api-zod/src/generated/types/updateMenuItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ethiopian Restaurant API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateMenuItemBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  nameAmharic?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  isVegan?: boolean | null;
  /** @nullable */
  isVegetarian?: boolean | null;
  /** @nullable */
  isGlutenFree?: boolean | null;
  /** @nullable */
  isAvailable?: boolean | null;
  /** @nullable */
  imageUrl?: string | null;
}